+++ src/write_config.R
@@ -2,19 +2,23 @@ write_config <- function(noro_nowcast_dates = NULL,
                          measles_nowcast_dates = NULL) {
   norovirus_url <- "https://raw.githubusercontent.com/jonathonmellor/norovirus-nowcast/refs/heads/main/outputs/data/cases_with_noise.csv" # nolint
   if (is.null(noro_nowcast_dates)) {
-    noro_nowcast_dates <- seq(
-      from = ymd("2013-05-05"),
-      to = ymd("2014-02-25"),
-      by = "day"
+    noro_nowcast_dates <- as.character(
+      seq(
+        from = ymd("2023-11-05"),
+        to = ymd("2024-03-10"),
+        by = "week"
+      )
     )
   }
 
   measles_url <- "https://raw.githubusercontent.com/kassteele/Nowcasting/refs/heads/master/data/measles_NL_2013_2014.dat" # nolint
   if (is.null(measles_nowcast_dates)) {
-    measles_nowcast_dates <- seq(
-      from = ymd("2023-11-05"),
-      to = ymd("2024-03-10"),
-      by = "week"
+    measles_nowcast_dates <- as.character(
+      seq(
+        from = ymd("2013-05-05"),
+        to = ymd("2014-02-25"),
+        by = "day"
+      )
     )
   }
 
@@ -26,7 +30,8 @@ write_config <- function(noro_nowcast_dates = NULL,
       n_history_delay = 42,
       n_history_uncertainty = 10,
       borrow_delay = FALSE,
-      borrow_uncertainty = FALSE
+      borrow_uncertainty = FALSE,
+      days_to_eval = 7
     ),
     measles = list(
       url = measles_url,
@@ -37,7 +42,9 @@ write_config <- function(noro_nowcast_dates = NULL,
       borrow_delay = FALSE,
       borrow_uncertainty = FALSE
     ),
-    n_draws = 100
+    n_draws = 100,
+    Hub_quantiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
+    eval_timeframe = 80
   )
 
   yaml::write_yaml(config,