+++ NEWS.md
@@ -1,5 +1,6 @@
 # baselinenowcastpaper 0.0.0.1000
 
+-   Generate quantiles for each nowcast and create plots to overlay nowcasts and final data
 -   Set-up a generation of a single nowcast using targets
 -   Added pipeline outline
 -   Added repository skeleton.